requirements.txt
python-docx
pandarallel
gensim==4.0.1
matplotlib==3.4.3
nltk==3.6.2
numpy==1.21.2
pandas==1.3.2
scikit_learn==0.24.2
#torch==1.9.0
tqdm==4.62.2
